[AIR] Move integration logging callbacks to AIR
