+++ python/ray/tune/examples/mlflow_example.py
@@ -8,7 +8,7 @@ import time
 import mlflow
 
 from ray import tune
-from ray.tune.integration.mlflow import MLflowLoggerCallback, mlflow_mixin
+from ray.air.callbacks.mlflow import MLflowLoggerCallback, mlflow_mixin
 
 
 def evaluation_fn(step, width, height):
