+++ python/ray/tune/examples/wandb_example.py
@@ -7,8 +7,8 @@ import wandb
 
 from ray import tune
 from ray.tune import Trainable
+from ray.air.callbacks.wandb import WandbLoggerCallback
 from ray.tune.integration.wandb import (
-    WandbLoggerCallback,
     WandbTrainableMixin,
     wandb_mixin,
 )
